BACKWARD-INCOMPATIBLE: Removed ``QueuedSearchIndex`` in favor of ``QueuedSignalProcessor``.
This brings ``queued_search`` back into compatibility with Haystack master.
