+++ queued_search/__init__.py
@@ -1,17 +1,2 @@
-from django.conf import settings
-
-
 __author__ = 'Daniel Lindsley'
 __version__ = (2, 0, 0, 'alpha')
-
-
-def get_queue_name():
-    """
-    Standized way to fetch the queue name.
-    
-    Can be overridden by specifying ``SEARCH_QUEUE_NAME`` in your settings.
-    
-    Given that the queue name is used in disparate places, this is primarily
-    for sanity.
-    """
-    return getattr(settings, 'SEARCH_QUEUE_NAME', 'haystack_search_queue')
