--- tests/notes/search_sites.py
@@ -1,2 +1,0 @@
-import haystack
-haystack.autodiscover()
