+++ tests/notes/tests.py
@@ -4,14 +4,14 @@ from django.core.management import call_command
 from django.test import TestCase
 from haystack import connections
 from haystack.query import SearchQuerySet
-from queued_search import get_queue_name
 from queued_search.management.commands.process_search_queue import Command as ProcessSearchQueueCommand
+from queued_search.utils import get_queue_name
 from notes.models import Note
 
 
 class AssertableHandler(logging.Handler):
     stowed_messages = []
-    
+
     def emit(self, record):
         AssertableHandler.stowed_messages.append(record.getMessage())
 
@@ -23,60 +23,60 @@ logging.getLogger('queued_search').addHandler(assertable)
 class QueuedSearchIndexTestCase(TestCase):
     def setUp(self):
         super(QueuedSearchIndexTestCase, self).setUp()
-        
+
         # Nuke the queue.
         queues.delete_queue(get_queue_name())
-        
+
         # Nuke the index.
         call_command('clear_index', interactive=False, verbosity=0)
-        
+
         # Get a queue connection so we can poke at it.
         self.queue = queues.Queue(get_queue_name())
-    
+
     def test_update(self):
         self.assertEqual(len(self.queue), 0)
-        
+
         note1 = Note.objects.create(
             title='A test note',
             content='Because everyone loves test data.',
             author='Daniel'
         )
-        
+
         self.assertEqual(len(self.queue), 1)
-        
+
         note2 = Note.objects.create(
             title='Another test note',
             content='More test data.',
             author='Daniel'
         )
-        
+
         self.assertEqual(len(self.queue), 2)
-        
+
         note3 = Note.objects.create(
             title='Final test note',
             content='The test data. All done.',
             author='Joe'
         )
-        
+
         self.assertEqual(len(self.queue), 3)
-        
+
         note3.title = 'Final test note FOR REAL'
         note3.save()
-        
+
         self.assertEqual(len(self.queue), 4)
-        
+
         # Pull the whole queue.
         messages = []
-        
+
         try:
             while True:
                 messages.append(self.queue.read())
         except QueueException:
             # We're out of queued bits.
             pass
-        
+
         self.assertEqual(messages, [u'update:notes.note.1', u'update:notes.note.2', u'update:notes.note.3', u'update:notes.note.3'])
-    
+
     def test_delete(self):
         note1 = Note.objects.create(
             title='A test note',
@@ -93,11 +93,11 @@ class QueuedSearchIndexTestCase(TestCase):
             content='The test data. All done.',
             author='Joe'
         )
-        
+
         # Dump the queue in preparation for the deletes.
         queues.delete_queue(get_queue_name())
         self.queue = queues.Queue(get_queue_name())
-        
+
         self.assertEqual(len(self.queue), 0)
         note1.delete()
         self.assertEqual(len(self.queue), 1)
@@ -105,158 +105,158 @@ class QueuedSearchIndexTestCase(TestCase):
         self.assertEqual(len(self.queue), 2)
         note3.delete()
         self.assertEqual(len(self.queue), 3)
-        
+
         # Pull the whole queue.
         messages = []
-        
+
         try:
             while True:
                 messages.append(self.queue.read())
         except QueueException:
             # We're out of queued bits.
             pass
-        
+
         self.assertEqual(messages, [u'delete:notes.note.1', u'delete:notes.note.2', u'delete:notes.note.3'])
-    
+
     def test_complex(self):
         self.assertEqual(len(self.queue), 0)
-        
+
         note1 = Note.objects.create(
             title='A test note',
             content='Because everyone loves test data.',
             author='Daniel'
         )
-        
+
         self.assertEqual(len(self.queue), 1)
-        
+
         note2 = Note.objects.create(
             title='Another test note',
             content='More test data.',
             author='Daniel'
         )
-        
+
         self.assertEqual(len(self.queue), 2)
-        
+
         note1.delete()
         self.assertEqual(len(self.queue), 3)
-        
+
         note3 = Note.objects.create(
             title='Final test note',
             content='The test data. All done.',
             author='Joe'
         )
-        
+
         self.assertEqual(len(self.queue), 4)
-        
+
         note3.title = 'Final test note FOR REAL'
         note3.save()
-        
+
         self.assertEqual(len(self.queue), 5)
-        
+
         note3.delete()
         self.assertEqual(len(self.queue), 6)
-        
+
         # Pull the whole queue.
         messages = []
-        
+
         try:
             while True:
                 messages.append(self.queue.read())
         except QueueException:
             # We're out of queued bits.
             pass
-        
+
         self.assertEqual(messages, [u'update:notes.note.1', u'update:notes.note.2', u'delete:notes.note.1', u'update:notes.note.3', u'update:notes.note.3', u'delete:notes.note.3'])
 
 
 class ProcessSearchQueueTestCase(TestCase):
     def setUp(self):
         super(ProcessSearchQueueTestCase, self).setUp()
-        
+
         # Nuke the queue.
         queues.delete_queue(get_queue_name())
-        
+
         # Nuke the index.
         call_command('clear_index', interactive=False, verbosity=0)
-        
+
         # Get a queue connection so we can poke at it.
         self.queue = queues.Queue(get_queue_name())
-        
+
         # Clear out and capture log messages.
         AssertableHandler.stowed_messages = []
-        
+
         self.psqc = ProcessSearchQueueCommand()
-    
+
     def test_process_mesage(self):
         self.assertEqual(self.psqc.actions, {'update': set([]), 'delete': set([])})
-        
+
         self.psqc.process_message('update:notes.note.1')
         self.assertEqual(self.psqc.actions, {'update': set(['notes.note.1']), 'delete': set([])})
-        
+
         self.psqc.process_message('delete:notes.note.2')
         self.assertEqual(self.psqc.actions, {'update': set(['notes.note.1']), 'delete': set(['notes.note.2'])})
-        
+
         self.psqc.process_message('update:notes.note.2')
         self.assertEqual(self.psqc.actions, {'update': set(['notes.note.1', 'notes.note.2']), 'delete': set([])})
-        
+
         self.psqc.process_message('delete:notes.note.1')
         self.assertEqual(self.psqc.actions, {'update': set(['notes.note.2']), 'delete': set(['notes.note.1'])})
-        
+
         self.psqc.process_message('wtfmate:notes.note.1')
         self.assertEqual(self.psqc.actions, {'update': set(['notes.note.2']), 'delete': set(['notes.note.1'])})
-        
+
         self.psqc.process_message('just plain wrong')
         self.assertEqual(self.psqc.actions, {'update': set(['notes.note.2']), 'delete': set(['notes.note.1'])})
-    
+
     def test_split_obj_identifier(self):
         self.assertEqual(self.psqc.split_obj_identifier('notes.note.1'), ('notes.note', '1'))
         self.assertEqual(self.psqc.split_obj_identifier('myproject.notes.note.73'), ('myproject.notes.note', '73'))
         self.assertEqual(self.psqc.split_obj_identifier('wtfmate.1'), ('wtfmate', '1'))
         self.assertEqual(self.psqc.split_obj_identifier('wtfmate'), (None, None))
-    
+
     def test_processing(self):
         self.assertEqual(len(self.queue), 0)
-        
+
         note1 = Note.objects.create(
             title='A test note',
             content='Because everyone loves test data.',
             author='Daniel'
         )
-        
+
         self.assertEqual(len(self.queue), 1)
-        
+
         note2 = Note.objects.create(
             title='Another test note',
             content='More test data.',
             author='Daniel'
         )
-        
+
         self.assertEqual(len(self.queue), 2)
-        
+
         note1.delete()
         self.assertEqual(len(self.queue), 3)
-        
+
         note3 = Note.objects.create(
             title='Final test note',
             content='The test data. All done.',
             author='Joe'
         )
-        
+
         self.assertEqual(len(self.queue), 4)
-        
+
         note3.title = 'Final test note FOR REAL'
         note3.save()
-        
+
         self.assertEqual(len(self.queue), 5)
-        
+
         note3.delete()
         self.assertEqual(len(self.queue), 6)
-        
+
         self.assertEqual(AssertableHandler.stowed_messages, [])
-        
+
         # Call the command.
         call_command('process_search_queue')
-        
+
         self.assertEqual(AssertableHandler.stowed_messages, [
             'Starting to process the queue.',
             u"Processing message 'update:notes.note.1'...",
@@ -285,24 +285,24 @@ class ProcessSearchQueueTestCase(TestCase):
             'Processing complete.'
         ])
         self.assertEqual(SearchQuerySet().all().count(), 1)
-    
+
     def test_requeuing(self):
         self.assertEqual(len(self.queue), 0)
-        
+
         note1 = Note.objects.create(
             title='A test note',
             content='Because everyone loves test data.',
             author='Daniel'
         )
-        
+
         self.assertEqual(len(self.queue), 1)
-        
+
         # Write a failed message.
         self.queue.write('update:notes.note.abc')
         self.assertEqual(len(self.queue), 2)
-        
+
         self.assertEqual(AssertableHandler.stowed_messages, [])
-        
+
         try:
             # Call the command, which will fail.
             call_command('process_search_queue')
@@ -311,22 +311,22 @@ class ProcessSearchQueueTestCase(TestCase):
             # We don't care that it failed. We just want to examine the state
             # of things afterward.
             pass
-        
+
         self.assertEqual(len(self.queue), 2)
-        
+
         # Pull the whole queue.
         messages = []
-        
+
         try:
             while True:
                 messages.append(self.queue.read())
         except QueueException:
             # We're out of queued bits.
             pass
-        
+
         self.assertEqual(messages, [u'update:notes.note.1', 'update:notes.note.abc'])
         self.assertEqual(len(self.queue), 0)
-        
+
         self.assertEqual(AssertableHandler.stowed_messages, [
             'Starting to process the queue.',
             u"Processing message 'update:notes.note.1'...",
@@ -340,34 +340,34 @@ class ProcessSearchQueueTestCase(TestCase):
             'Requeuing unprocessed messages.',
             'Requeued 2 updates and 0 deletes.'
         ])
-        
+
         # Start over.
         note1 = Note.objects.create(
             title='A test note',
             content='Because everyone loves test data.',
             author='Daniel'
         )
-        
+
         self.assertEqual(len(self.queue), 1)
-        
+
         note2 = Note.objects.create(
             title='Another test note',
             content='Because everyone loves test data.',
             author='Daniel'
         )
-        
+
         self.assertEqual(len(self.queue), 2)
-        
+
         # Now delete it.
         note2.delete()
-        
+
         # Write a failed message.
         self.queue.write('delete:notes.note.abc')
         self.assertEqual(len(self.queue), 4)
-        
+
         AssertableHandler.stowed_messages = []
         self.assertEqual(AssertableHandler.stowed_messages, [])
-        
+
         try:
             # Call the command, which will fail again.
             call_command('process_search_queue')
@@ -376,23 +376,23 @@ class ProcessSearchQueueTestCase(TestCase):
             # We don't care that it failed. We just want to examine the state
             # of things afterward.
             pass
-        
+
         # Everything but the bad bit of data should have processed.
         self.assertEqual(len(self.queue), 1)
-        
+
         # Pull the whole queue.
         messages = []
-        
+
         try:
             while True:
                 messages.append(self.queue.read())
         except QueueException:
             # We're out of queued bits.
             pass
-        
+
         self.assertEqual(messages, ['delete:notes.note.abc'])
         self.assertEqual(len(self.queue), 0)
-        
+
         self.assertEqual(AssertableHandler.stowed_messages, [
             'Starting to process the queue.',
             u"Processing message 'update:notes.note.2'...",
